refactor preprocessing, io routines to radiation package with better names
