+++ .pre-commit-config.yaml
@@ -44,6 +44,7 @@ repos:
         workflows/prognostic_c48_run/tests/.+ |
         external/fv3fit/fv3fit/.+ |
         external/loaders/loaders/.+ |
+        external/radiation/radiation/.+ |
         workflows/diagnostics/fv3net/diagnostics/offline/.+ |
         workflows/diagnostics/fv3net/diagnostics/prognostic_run/.+
         )$
