+++ external/radiation/radiation/__init__.py
@@ -1,5 +1,5 @@
 from .radiation_driver import RadiationDriver
-from .variables_to_read import vars_dict
-from . import getdata
+from . import io, preprocessing
+from .config import get_rad_config, LOOKUP_DATA_PATH, FORCING_DATA_PATH
 from .radsw import ngptsw as NGPTSW
 from .radlw import ngptlw as NGPTLW
