--- workflows/prognostic_c48_run/runtime/steppers/radiation/__init__.py
@@ -1,1 +1,0 @@
-from ._stepper import RadiationStepper, RadiationConfig
